fix deprecated methods::as(A, "dgCMatrix") in JANE.R
Use  A <- methods::as(A, "generalMatrix") instead but need to also first convert A to sparse matrix if A is dense matrix otherwise it will fail. Specfically, for a dense matrix A as(A,  "generalMatrix") will just return dense matrix. So need to first convert dense matrix to sparse matrix